src/tests/test_tutorial_resources.py
```python
import unittest
from src import tutorial_resources

class TestTutorialResources(unittest.TestCase):

    def setUp(self):
        self.tutorial_resources = tutorial_resources.TutorialResources()
        self.userProfile = {
            "username": "test_user",
            "email": "test_user@example.com",
            "password": "test_password",
            "dance_styles": ["contemporary", "hip-hop"],
            "subscriptionStatus": "free"
        }

    def test_access_tutorial_resource(self):
        # Test if the user can access the tutorial resources
        tutorial_resource_id = "tutorial_1"
        result = self.tutorial_resources.accessTutorialResource(self.userProfile, tutorial_resource_id)
        self.assertEqual(result, True)

    def test_get_tutorial_resources(self):
        # Test if the user can get the list of tutorial resources
        result = self.tutorial_resources.getTutorialResources(self.userProfile)
        self.assertIsInstance(result, list)

    def test_get_tutorial_resource_by_id(self):
        # Test if the user can get a specific tutorial resource by id
        tutorial_resource_id = "tutorial_1"
        result = self.tutorial_resources.getTutorialResourceById(self.userProfile, tutorial_resource_id)
        self.assertIsInstance(result, dict)

    def test_get_tutorial_resources_by_dance_style(self):
        # Test if the user can get tutorial resources by dance style
        dance_style = "contemporary"
        result = self.tutorial_resources.getTutorialResourcesByDanceStyle(self.userProfile, dance_style)
        self.assertIsInstance(result, list)

if __name__ == '__main__':
    unittest.main()
```